Add files via upload
This is the result of data_words before step #9 from https://www.machinelearningplus.com/nlp/topic-modeling-gensim-python/.

I can't install spacy to lemmatize and I'm not really sure how I would print out the result from #10 so I just left it at this.

I wrote the file using csv.writerows(data_words) so maybe you should read it in with csv.reader. data_words is a list of each row in the output file which is itself a list of the individual words if that makes any sense.

One thing I noticed that I thought was interesting was that all the numbers were removed, not sure how that will turn out but we'll see.